gdb/testsuite: split gdb.dwarf2/macro-source-path.exp
Because it runs so many variations, the test
gdb.dwarf2/macro-source-path.exp takes about 2:40 minutes to run for me,
in a non-optimized build.  These days I often run all tests under
gdb.dwarf2, as a sanity test for my changes, and so I often have to wait
for this test to complete.

Split the test, to allow it to complete faster when running the
testsuite in parallel.  After this patch, running all the
gdb.dwarf2/macro-source-path-*.exp tests in parallel takes me about 1
minute.  It's more that I would expect, I would expect the time to be
divided by nearly 5, but it's already better than what we have now.

Change-Id: I07e4e1f234cf57d9b0c1c027f08061615714a4d5
Acked-By: Tom de Vries <tdevries@suse.de>